Update template with course name for e298dh_test
"e298dh_test" is the name of the course that we (FAS) are looking to test on FAS OnDemand. As such, we have created a new branch, also called e298dh_test, wherein we've made this update to the /template/before.sh.erb script. Our understanding is that each course will have its own branch.